Canonicalize audio readiness return types

+++ include/main/audio/snd_core.h
@@ -5,7 +5,7 @@
 
 void sndQuit(void);
 void sndSetMaxVoices(u8 valueA, u8 valueB);
-u8 sndIsInstalled(void);
+u32 sndIsInstalled(void);
 void salApplyMatrix(f32 *matrix, f32 *vec, f32 *out);
 f32 salNormalizeVector(f32 *v);
 void inpSetGlobalMIDIDirtyFlag(u8 index, u8 group, u32 flags);
